+++ dash/dash.py
@@ -352,16 +352,21 @@ class Dash(object):
         ).format(json.dumps(self._config()))
 
     def _generate_meta_html(self):
+        has_ie_compat = any(
+            x.get('http-equiv', '') == 'X-UA-Compatible'
+            for x in self._meta_tags)
         has_charset = any('charset' in x for x in self._meta_tags)
 
         tags = []
+        if not has_ie_compat:
+            tags.append('<meta equiv="X-UA-Compatible" content="IE=edge">')
         if not has_charset:
             tags.append('<meta charset="UTF-8"/>')
 
         tags = tags + [
             _format_tag('meta', x, opened=True) for x in self._meta_tags
         ]
-
+        
         return '\n      '.join(tags)
 
     # Serve the JS bundles for each package
@@ -634,6 +639,110 @@ class Dash(object):
                 output.component_id,
                 output.component_property).replace('    ', ''))
 
+    def _validate_callback_output(self, output_value, output):
+        valid = [str, dict, int, float, type(None), Component]
+
+        def _raise_invalid(bad_val, outer_val, bad_type, path, index=None,
+                           toplevel=False):
+            outer_id = "(id={:s})".format(outer_val.id) \
+                        if getattr(outer_val, 'id', False) else ''
+            outer_type = type(outer_val).__name__
+            raise exceptions.InvalidCallbackReturnValue('''
+            The callback for property `{property:s}` of component `{id:s}`
+            returned a {object:s} having type `{type:s}`
+            which is not JSON serializable.
+
+            {location_header:s}{location:s}
+            and has string representation
+            `{bad_val}`
+
+            In general, Dash properties can only be
+            dash components, strings, dictionaries, numbers, None,
+            or lists of those.
+            '''.format(
+                property=output.component_property,
+                id=output.component_id,
+                object='tree with one value' if not toplevel else 'value',
+                type=bad_type,
+                location_header=(
+                    'The value in question is located at'
+                    if not toplevel else
+                    '''The value in question is either the only value returned,
+                    or is in the top level of the returned list,'''
+                ),
+                location=(
+                    "\n" +
+                    ("[{:d}] {:s} {:s}".format(index, outer_type, outer_id)
+                     if index is not None
+                     else ('[*] ' + outer_type + ' ' + outer_id))
+                    + "\n" + path + "\n"
+                ) if not toplevel else '',
+                bad_val=bad_val).replace('    ', ''))
+
+        def _value_is_valid(val):
+            return (
+                # pylint: disable=unused-variable
+                any([isinstance(val, x) for x in valid]) or
+                type(val).__name__ == 'unicode'
+            )
+
+        def _validate_value(val, index=None):
+            # val is a Component
+            if isinstance(val, Component):
+                for p, j in val.traverse_with_paths():
+                    # check each component value in the tree
+                    if not _value_is_valid(j):
+                        _raise_invalid(
+                            bad_val=j,
+                            outer_val=val,
+                            bad_type=type(j).__name__,
+                            path=p,
+                            index=index
+                        )
+
+                    # Children that are not of type Component or
+                    # collections.MutableSequence not returned by traverse
+                    child = getattr(j, 'children', None)
+                    if not isinstance(child, collections.MutableSequence):
+                        if child and not _value_is_valid(child):
+                            _raise_invalid(
+                                bad_val=child,
+                                outer_val=val,
+                                bad_type=type(child).__name__,
+                                path=p + "\n" + "[*] " + type(child).__name__,
+                                index=index
+                            )
+
+                # Also check the child of val, as it will not be returned
+                child = getattr(val, 'children', None)
+                if not isinstance(child, collections.MutableSequence):
+                    if child and not _value_is_valid(child):
+                        _raise_invalid(
+                            bad_val=child,
+                            outer_val=val,
+                            bad_type=type(child).__name__,
+                            path=type(child).__name__,
+                            index=index
+                        )
+
+            # val is not a Component, but is at the top level of tree
+            else:
+                if not _value_is_valid(val):
+                    _raise_invalid(
+                        bad_val=val,
+                        outer_val=type(val).__name__,
+                        bad_type=type(val).__name__,
+                        path='',
+                        index=index,
+                        toplevel=True
+                    )
+
+        if isinstance(output_value, list):
+            for i, val in enumerate(output_value):
+                _validate_value(val, index=i)
+        else:
+            _validate_value(output_value)
+
     # TODO - Update nomenclature.
     # "Parents" and "Children" should refer to the DOM tree
     # and not the dependency tree.
@@ -680,9 +789,26 @@ class Dash(object):
                     }
                 }
 
+                try:
+                    jsonResponse = json.dumps(
+                        response,
+                        cls=plotly.utils.PlotlyJSONEncoder
+                    )
+                except TypeError:
+                    self._validate_callback_output(output_value, output)
+                    raise exceptions.InvalidCallbackReturnValue('''
+                    The callback for property `{property:s}`
+                    of component `{id:s}` returned a value
+                    which is not JSON serializable.
+
+                    In general, Dash properties can only be
+                    dash components, strings, dictionaries, numbers, None,
+                    or lists of those.
+                    '''.format(property=output.component_property,
+                               id=output.component_id))
+
                 return flask.Response(
-                    json.dumps(response,
-                               cls=plotly.utils.PlotlyJSONEncoder),
+                    jsonResponse,
                     mimetype='application/json'
                 )